+++ src/components/Table/TableReel.tsx
@@ -1,47 +1,15 @@
-import React, { HTMLAttributes, useEffect, useRef, useState } from 'react'
+import React, { HTMLAttributes, PropsWithChildren, useRef, useState } from 'react'
 import styled, { css } from 'styled-components'
 
-export type Props = {
-  children: React.ReactNode
-}
-type ElementProps = Omit<HTMLAttributes<HTMLDivElement>, keyof Props>
+import { useReelCells } from './useReelCells'
 
-export const TableReel: React.FC<Props & ElementProps> = ({ children, ...props }) => {
+type ElementProps = Omit<HTMLAttributes<HTMLDivElement>, keyof PropsWithChildren>
+
+export const TableReel: React.FC<PropsWithChildren & ElementProps> = ({ children, ...props }) => {
   const tableWrapperRef = useRef<HTMLDivElement>(null)
   const [showShadow, setShowShadow] = useState(false)
 
-  useEffect(() => {
-    const currentRef = tableWrapperRef.current
-
-    const handleScroll = () => {
-      if (currentRef) {
-        const stickyCells = currentRef.querySelectorAll('.fixedElement') || []
-        const scrollLeft = currentRef.scrollLeft
-        const maxScrollLeft = currentRef.scrollWidth - currentRef.clientWidth || 0
-
-        stickyCells.forEach((cell) => {
-          const shouldFix = maxScrollLeft > 0 && scrollLeft < maxScrollLeft
-
-          if (shouldFix) {
-            cell.classList.add('fixed')
-            setShowShadow(scrollLeft > 0)
-          } else {
-            cell.classList.remove('fixed')
-            setShowShadow(maxScrollLeft === 0 && scrollLeft === 0 ? false : true)
-          }
-        })
-      }
-    }
-    handleScroll()
-
-    window.addEventListener('resize', handleScroll)
-    currentRef?.addEventListener('scroll', handleScroll)
-
-    return () => {
-      window.removeEventListener('resize', handleScroll)
-      currentRef?.removeEventListener('scroll', handleScroll)
-    }
-  }, [tableWrapperRef, setShowShadow])
+  useReelCells(tableWrapperRef, setShowShadow)
 
   return (
     <Shadow showShadow={showShadow}>
@@ -58,16 +26,18 @@ type ReelShadow = {
 }
 
 export const reelShadow = ({ showShadow = true, direction = 'left' }: ReelShadow) => {
+  const shadowWidth = '12px' //影の横幅はpx指定で固定とする
+
   return `
     &::after {
       content: '';
       position: absolute;
       z-index: 0;
-      left: ${direction === 'left' ? '0' : '-12px'};
+      left: ${direction === 'left' ? '0' : `-${shadowWidth}`};
       top: 0;
-      width: 12px;
+      width: ${shadowWidth};
       pointer-events: none; /* 影の領域が広すぎるとクリッカブルエリアを侵食するので無効化 */
-      height: 100%;
+      inset-block: 0;
       background: linear-gradient(${
         direction === 'left' ? '90deg' : '-90deg'
       }, rgba(0, 0, 0, 0.2) 0%, rgba(0, 0, 0, 0) 100%);
